New version 1.17 - fixed all bugs.

+++ GooglePlay-ver.2/Sofbus24/src/bg/znestorov/sofbus24/station_database/FavouritesDataSource.java
@@ -13,6 +13,7 @@ import android.preference.PreferenceManager;
 import bg.znestorov.sofbus24.utils.Constants;
 import bg.znestorov.sofbus24.utils.TranslatorCyrillicToLatin;
 import bg.znestorov.sofbus24.utils.TranslatorLatinToCyrillic;
+import bg.znestorov.sofbus24.utils.Utils;
 
 public class FavouritesDataSource {
 
@@ -124,7 +125,10 @@ public class FavouritesDataSource {
 		String where = FavouritesSQLite.COLUMN_ID + " = ?";
 		String[] whereArgs = new String[] { String.valueOf(station.getId()) };
 
-		database.delete(FavouritesSQLite.TABLE_FAVOURITES, where, whereArgs);
+		if (!database.isOpen()) {
+			this.open();
+			database.delete(FavouritesSQLite.TABLE_FAVOURITES, where, whereArgs);
+		}
 	}
 
 	// Update station from the database
@@ -195,6 +199,10 @@ public class FavouritesDataSource {
 	private GPSStation cursorToStation(Cursor cursor) {
 		GPSStation station = new GPSStation();
 
+		// Format station number to be always 4 digits (or more)
+		String stationNumber = cursor.getString(0);
+		stationNumber = Utils.formatNumberOfDigits(stationNumber, 4);
+
 		// Check if have to translate the station name
 		String stationName = cursor.getString(1);
 		if (!"bg".equals(language)) {
@@ -202,7 +210,7 @@ public class FavouritesDataSource {
 		}
 
 		// Getting all columns of the row and setting them to a Station object
-		station.setId(cursor.getString(0));
+		station.setId(stationNumber);
 		station.setName(stationName);
 		station.setLat(cursor.getString(2));
 		station.setLon(cursor.getString(3));
